+++ backend/libraries/utils/src/cycles/check_cycles_balance.rs
@@ -1,5 +1,4 @@
 use crate::consts::CYCLES_REQUIRED_FOR_UPGRADE;
-use crate::memory;
 use canister_client_macros::generate_c2c_call;
 use std::cmp::max;
 use tracing::error;
@@ -23,7 +22,7 @@ fn should_notify() -> bool {
 }
 
 fn get_approx_freeze_threshold_cycles() -> Cycles {
-    let approx_memory_usage = memory::used();
+    let approx_memory_usage = crate::memory::used();
 
     let one_gib = 1 << 30;